client/requirements.txt
attrs
base58
bitarray
certifi
chardet
cytoolz
docker
eth-abi
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing
eth-utils
hexbytes
idna
importlib-metadata
importlib-resources
jsonschema
kafka-python
lru-dict
multiaddr
netaddr
parsimonious
protobuf
pycryptodome
pyrsistent
requests
rlp
six
toolz
typing-extensions
urllib3
varint
web3
websockets
zipp
ipfshttpclient